lws_struct: conditional prototypes make trouble with visibility
Although the code is correct and the symbols should be defined, making some
prototypes conditional on cmake defines breaks their visibility when the
library is built.

They're conditional here to get around use of sqlite3 type in the prototypes
where we may not be including sqlite3 headers.  Replace the conditional with
a grauitous typedef as a forward ref, since it's only referred to via
pointer types.