sspc-proxy-fragmented-dsh-must-use-fragmented-flags
We compute the refragmented flags when cutting up large client serialized
payload blocks.  But we had a bug where we didn't actually apply it and
applied the original client flags on the fragments.

That causes a crisis because EOM is used to mark end of post body and
complete the transaction, that is then happening on the first fragment.

This one-liner corrects it to use the computed, refragmented flags on the
dsh fragments and eliminate the problem.